Add more options for button styling (#1158)
This change adds more variants for styling buttons. This change is aimed
to ultimately help provide a better UX.

Buttons currently default to 'minimal' which styles the text the same
color as the surrounding text, which makes them difficult to distinguish
from the surrounding text.

This change adds the following variants and intents:

Variants (style):
- Filled
- Outlined
- Minimal

Intents (color):
- Primary
- Danger
- Success
- Warning

This change keeps the default variant as 'minimal', which matches what
the vast majority of buttons are in the UI currently, as to not
introduce too many changes in this change. However, folks should really
prefer using 'filled' or 'outlined' in the future to improve UX.