Retain DOM state between page changes (#1064)
Retain DOM state when switching pages

Currently the DOM state of a page (such as scroll position) is lost when
switching between pages.

This is a problem for example when switching from the timeline page to
the query page and back again, the scroll position within the timeline
page is lost, which can be quite annoying to recover.

This CL modifies the PageManager to keep all previously viewed pages in
the DOM, but makes all inactive pages invisible with `display: none`. As
an optimization, we avoid rendering inactive pages as well by caching
the vnodes of those pages.

Bug: [404946830](https://issuetracker.google.com/issues/404946830)